+++ packages/components/src/assets/icons/index.ts
@@ -1,2 +1,1 @@
-export { default as CheckCircleIcon } from "./check-circle.svg";
 export { default as CheckmarkIcon } from "./checkmark.svg";
